Renamed symbtrdataextractor/symbtr.py -> symbtrdataextractor/ScoreProcessor.py
