+++ symbtrdataextractor/PhraseExtractor.py
@@ -1,4 +1,4 @@
-from symbtr import ScoreProcessor
+from ScoreProcessor import ScoreProcessor
 from structure_label import label_structures, get_symbtr_labels
