+++ symbtrdataextractor/ScoreProcessor.py
